backend/requirements.txt
fastapi
uvicorn
pydantic-settings
google-auth-oauthlib
google-api-python-client
langgraph
langchain-core
langchain-openai
python-dotenv
llama-index
llama-index-vector-stores-chroma
chromadb
python-multipart
pypdf
llama-index-embeddings-gemini
llama-index-llms-gemini
fpdf
langgraph-checkpoint
